cleanup do teste

+++ src/test/generate.sh
@@ -1,4 +1,6 @@
-rm -rf mvnthings
+if [ -e mvnthings ]; then
+	rm -rf mvnthings
+fi
 mvn archetype:generate -DgroupId=com.teste -DartifactId=mvnthings -DarchetypeArtifactId=things-app-archetype -DarchetypeGroupId=org.crypthing -DarchetypeCatalog=local -DinteractiveMode=false
 cd mvnthings
 git init
@@ -6,4 +8,7 @@ git add .
 git commit -a -m "First commit"
 mvn package
 cd ..
+if [ -e mvnthings ]; then
+	rm -rf mvnthings
+fi
